First commit for sub20 project
This is a project for which the frontend is written in React, and the Back in
Rust's Rocket Framework.

I hope to have a working chatroom in which two people (after proper login/authentication)
are able to play a fun game together.

I hope to deploy the application to a cloud service, and assign it a custom domain when I do so.