Improved homepage UI and added some MVC separation
All login cookie related logic is now in Login Context.
Login Context provides a cookie value which we can read, and the ability
to set its value.

We can read this value in Home page to learn whether the user is logged in or not,
and display the buttons accordingly.

The error page can also use the data from this context, as receiving a 401
error indicates an invalid or missing token, so we treat this as needing to sign the user out if needed, and clear
their cookie.